numero de conta automatico, findall de conta corrente, enums para car…

+++ src/main/java/com/wipro/projetofinal/dto/AccountDTO.java
@@ -17,7 +17,7 @@ import com.wipro.projetofinal.entities.User;
 public class AccountDTO {
 
 	
-	protected Integer accountNumber = 0;
+	protected Long accountNumber = 0L;
 	protected Double balance;
 	
     protected Customer customer;
@@ -36,7 +36,7 @@ public class AccountDTO {
 	}
 
 
-	public Integer getAccountNumber() {
+	public Long getAccountNumber() {
 		return accountNumber;
 	}
